pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gha-doc"
version = "0.1.0"
description = "GitHub Actions Documentation Generator"
authors = [
    {name = "HafslundEcoVannkraft", email = "devops@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyyaml>=6.0",
    "markdown>=3.3",
    "python-frontmatter>=1.0.0",
    "diagrams>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
ai = [
    "openai>=1.0.0",
]

[tool.setuptools]
packages = ["src"]

[project.scripts]
gha-doc = "src.main:main"
